net, localnet: move liblocalnet to shared network libs
Move localnet helpers and constants from tests/network/localnet/liblocalnet.py
to tests/network/libs/localnet.py — the module is already imported by l2_bridge
and upgrade tests, so it belongs in the shared location.
