+++ tests/network/localnet/rhel9_rhel10_cluster/conftest.py
@@ -17,7 +17,7 @@ from libs.vm.vm import BaseVirtualMachine
 from tests.network.libs import cloudinit
 from tests.network.libs import cluster_user_defined_network as libcudn
 from tests.network.libs.cloudinit import EthernetDevice
-from tests.network.localnet.liblocalnet import (
+from tests.network.libs.localnet import (
     GUEST_2ND_IFACE_NAME,
     LOCALNET_BR_EX_INTERFACE,
     localnet_cloudinit,
