+++ tests/network/localnet/test_ovs_bridge.py
@@ -9,7 +9,7 @@ import pytest
 from libs.net.ip import filter_link_local_addresses, have_same_ip_families
 from libs.net.traffic_generator import client_server_active_connection, is_tcp_connection
 from libs.net.vmspec import lookup_iface_status
-from tests.network.localnet.liblocalnet import (
+from tests.network.libs.localnet import (
     GUEST_1ST_IFACE_NAME,
     LINK_STATE_UP,
     LOCALNET_OVS_BRIDGE_INTERFACE,
